@@ -1,8 +1,8 @@
 package com.xpluscloud.moses_apc;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Created by Shirwen on 10/19/2017.
