@@ -4,8 +4,8 @@ import android.content.Context;
 import android.graphics.Color;
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
@@ -63,9 +63,9 @@ public class PDFViewActivity extends AppCompatActivity {
     }
 
     private void loadFromFile(String filename){
-        String root = Environment.getExternalStorageDirectory().toString();
-        File myDir = new File(root + "/ShellPromo");
-        File pdfFile = new File(root + "/ShellPromo/"+filename);
+        String root = ctx.getFilesDir().toString();
+        File myDir = new File(root + "/apc");
+        File pdfFile = new File(root + "/apc/"+filename);
         if(myDir.isDirectory() && pdfFile.isFile()) {
 
 
