@@ -2,7 +2,7 @@ package com.xpluscloud.moses_apc;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 public class StoretypeActivity extends Activity {
 
