@@ -8,9 +8,9 @@ import android.net.NetworkInfo;
 import android.os.AsyncTask;
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.view.Gravity;
 import android.view.LayoutInflater;
@@ -23,7 +23,6 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.xpluscloud.moses_apc.dbase.CustomerDbManager;
-import com.xpluscloud.moses_apc.dbase.PictureDbManager;
 import com.xpluscloud.moses_apc.dbase.SignatureDbManager;
 import com.xpluscloud.moses_apc.getset.Customer;
 
@@ -56,7 +55,7 @@ public class UploadSignatureActivity extends AppCompatActivity {
         	//Resources res = getResources(); 
 
         	uploadURL = getString(R.string.sig_file_upload);
-	        FilePath = Environment.getExternalStorageDirectory() + "/"
+	        FilePath = context.getFilesDir().toString() +"/"
 	  	            + getResources().getString(R.string.signatures_dir) + "/";
 	  	    
 	  	   // sFolder +=  folder +"/";
