@@ -11,7 +11,7 @@ import android.location.LocationManager;
 import android.os.BatteryManager;
 import android.os.Build;
 import android.provider.Settings;
-import android.support.v4.app.ActivityCompat;
+import androidx.core.app.ActivityCompat;
 import android.telephony.TelephonyManager;
 
 import java.util.ArrayList;
@@ -156,7 +156,7 @@ public final class Master {
     /*******************************
      * Outgoing  Commands ***********
      *******************************/
-    public final static int FOR_APPROVAL_SETTING = 0; //if this app is for approval status 1
+    public final static int FOR_APPROVAL_SETTING = 1; //if this app is for approval status 1
     public final static String CMD_REGISTER = "CMDREG1";
     public final static String CMD_INOUT = "CMDTIO"; //Time In Out
 
